Add CSS selector support in Selenium utils
Added a new method in `selenium_utils.py` to wait for an element to be located by CSS selector. Before this commit, we only had support for XPath selectors. With this addition, we can wait for an element located by either XPath or CSS selector, providing flexibility in handling different types of web elements.